Alterando nome de SucessModal para RequestResponseModal, e deixando ele um componente reutilizavel para mensagem de sucesso ou falha
